refactor: extract the function convertVideoToAudio

+++ src/index.ts
@@ -19,7 +19,7 @@ const watchFolder = (inputFolder: string) => {
   });
 };
 
-const handleNewFile = (filePath: string) => {
+const handleNewFile = async (filePath: string) => {
   const { name: fileName, ext: fileExt } = path.parse(filePath);
 
   const outputFilePath = path.join(".", outputFolder, fileName);
@@ -30,20 +30,38 @@ const handleNewFile = (filePath: string) => {
 
   const audioPath = path.join(outputFilePath, "audio.mp3");
 
-  ffmpeg(filePath)
-    .audioCodec("libmp3lame")
-    .save(audioPath)
-    .on("end", () => {
-      console.log("Audio file created", audioPath);
+  try {
+    await convertVideoToAudio(filePath, audioPath);
+    console.log("Audio file created", audioPath);
 
-      const videoOutputPath = path.join(outputFilePath, fileName + fileExt);
+    const videoOutputPath = path.join(outputFilePath, fileName + fileExt);
+
+    if (fs.existsSync(filePath)) {
       fs.renameSync(filePath, videoOutputPath);
+      console.log("Video file moved", videoOutputPath);
+    }
+  } catch (error) {
+    console.log("Error converting video to audio", error);
+  }
+};
 
-      console.log("Video file moved to output folder", videoOutputPath);
-    })
-    .on("error", (error) => {
-      console.log("Error converting video to audio", error);
-    });
+const convertVideoToAudio = (
+  inputPath: string,
+  outputPath: string
+): Promise<void> => {
+  return new Promise((resolve, reject) => {
+    ffmpeg(inputPath)
+      .audioCodec("libmp3lame")
+      .save(outputPath)
+      .on("end", () => {
+        console.log("Audio file created", outputPath);
+        resolve();
+      })
+      .on("error", (error) => {
+        console.log("Error converting video to audio", error);
+        reject();
+      });
+  });
 };
 
 watchFolder(inputFolder);
